Razorpay_Django/Razorpay_Django/requirements.txt
dj-rest-auth==2.1.3
Django==3.2.4
# django-allauth==0.44.0
django-allauth==0.44.0
django-cors-headers==3.7.0
django-jsonfield==1.4.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
razorpay==1.2.0
Pillow==8.2.0
PyJWT==2.1.0